Struct rusty_machine::data::transforms::shuffle::Shuffler

pub struct Shuffler<R: Rng> { /* fields omitted */ }

The Shuffler

Provides an implementation of Transformer which shuffles the input rows in place.

Methods

impl<R: Rng> Shuffler<R>

fn new(rng: R) -> Self

Construct a new Shuffler with given random number generator.

Examples

use rusty_machine::data::transforms::Transformer;
use rusty_machine::data::transforms::shuffle::Shuffler;
use rand::{StdRng, SeedableRng};

// We can create a seeded rng
let rng = StdRng::from_seed(&[1, 2, 3]);

let shuffler = Shuffler::new(rng);