Struct rusty_machine::data::transforms::standardize::Standardizer

pub struct Standardizer<T: Float> {
    // some fields omitted
}

The Standardizer

The Standardizer provides an implementation of Transformer which allows us to transform the input data to have a new mean and standard deviation.

See the module description for more information.

Methods

impl<T: Float> Standardizer<T>

fn new(mean: T, stdev: T) -> Standardizer<T>

Constructs a new Standardizer with the given mean and variance

Examples

use rusty_machine::data::transforms::Standardizer;

// Constructs a new `Standardizer` which will give the data
// mean `0` and standard deviation `2`.
let transformer = Standardizer::new(0.0, 2.0);