Function rusty_machine::datasets::iris::load

pub fn load() -> Dataset<Matrix<f64>, Vector<usize>>

Load iris dataset.

The data set contains 3 classes of 50 instances each, where each class refers to a type of iris plant.

Attribute Information

Data

Matrix<f64> contains following columns.

sepal length in cm

sepal width in cm

petal length in cm

petal width in cm

Target

Vector<usize> contains numbers corresponding to iris species:

0: Iris Setosa

1: Iris Versicolour

2: Iris Virginica

Lichman, M. (2013). UCI Machine Learning Repository [http://archive.ics.uci.edu/ml]. Irvine, CA: University of California, School of Information and Computer Science.