Struct rusty_machine::learning::dbscan::DBSCAN

pub struct DBSCAN { /* fields omitted */ }

DBSCAN Model

Implements clustering using the DBSCAN algorithm via the UnSupModel trait.

Methods

impl DBSCAN

fn new(eps: f64, min_points: usize) -> DBSCAN

Create a new DBSCAN model with a given distance episilon and minimum points per cluster.

fn set_predictive(&mut self, predictive: bool)

Set predictive to true if the model is to be used to classify future points.

If the model is set as predictive then the input data will be cloned during training.

fn clusters(&self) -> Option<&Vector<Option<usize>>>

Return an Option pointing to the model clusters.