Enum rusty_machine::learning::error::ErrorKind

pub enum ErrorKind {
    InvalidParameters,
    InvalidData,
}

Types of errors produced in the learning module.

List intended to grow and so you should be wary of matching against explicitly.

Variants
InvalidParameters

The parameters used to define the model are not valid.

InvalidData

The input data to the model is not valid.