Struct rusty_machine::learning::glm::GenLinearModel

pub struct GenLinearModel<C: Criterion> {
    // some fields omitted
}

The Generalized Linear Model

The model is generic over a Criterion which specifies the distribution family and the link function.

Methods

impl<C: Criterion> GenLinearModel<C>

fn new(criterion: C) -> GenLinearModel<C>

Constructs a new Generalized Linear Model.

Takes a Criterion which fully specifies the family and the link function used by the GLM.

use rusty_machine::learning::glm::GenLinearModel;
use rusty_machine::learning::glm::Bernoulli;

let glm = GenLinearModel::new(Bernoulli);