Struct rusty_machine::learning::gp::GaussianProcess

pub struct GaussianProcess<T: Kernel, U: MeanFunc> {
    pub noise: f64,
    // some fields omitted
}

Gaussian Process struct

Gaussian process with generic kernel and deterministic mean function. Can be used for gaussian process regression with noise. Currently does not support classification.

Fields

noise

The observation noise of the GP.

Methods

impl<T: Kernel, U: MeanFunc> GaussianProcess<T, U>

fn new(ker: T, mean: U, noise: f64) -> GaussianProcess<T, U>

Construct a new Gaussian Process.

Examples

use rusty_machine::learning::gp;
use rusty_machine::learning::toolkit::kernel;

let ker = kernel::SquaredExp::default();
let mean = gp::ConstMean::default();
let gaussp = gp::GaussianProcess::new(ker, mean, 1e-3f64);

impl<T: Kernel, U: MeanFunc> GaussianProcess<T, U>

fn get_posterior(&self, data: &Matrix<f64>) -> (Matrix<f64>, Matrix<f64>)

Compute the posterior distribution [UNSTABLE]

Requires the model to be trained first. Note that this is a messy compromise as GPs do not fit the SupModel trait as is currently implemented.