Struct rusty_machine::learning::naive_bayes::NaiveBayes

pub struct NaiveBayes<T: Distribution> {
    // some fields omitted
}

The Naive Bayes model.

Methods

impl<T: Distribution> NaiveBayes<T>

fn new() -> NaiveBayes<T>

Create a new NaiveBayes model from a given distribution.

Examples

use rusty_machine::learning::naive_bayes::{NaiveBayes, Gaussian};

// Create a new Gaussian Naive Bayes model.
let _ = NaiveBayes::<Gaussian>::new();

fn cluster_count(&self) -> &Option<usize>

Get the cluster count for this model.

Returns an option which is None until the model has been trained.

fn class_prior(&self) -> &Option<Vec<f64>>

Get the class prior distribution for this model.

Returns an option which is None until the model has been trained.

fn distr(&self) -> &Option<T>

Get the distribution for this model.

Returns an option which is None until the model has been trained.

impl<T: Distribution> NaiveBayes<T>

fn get_log_probs(&self, inputs: &Matrix<f64>) -> Matrix<f64>

Get the log-probabilities per class for each input.