Module rusty_machine::learning::nnet

Neural Network module

Contains implementation of simple feed forward neural network.

Usage

use rusty_machine::learning::nnet::NeuralNet;
use rusty_machine::linalg::matrix::Matrix;
use rusty_machine::learning::SupModel;

let inputs = Matrix::new(5,3, vec![1.,1.,1.,2.,2.,2.,3.,3.,3.,
                                4.,4.,4.,5.,5.,5.,]);
let targets = Matrix::new(5,3, vec![1.,0.,0.,0.,1.,0.,0.,0.,1.,
                                    0.,0.,1.,0.,0.,1.]);

let layers = &[3,5,11,7,3];
let mut model = NeuralNet::default(layers);

model.train(&inputs, &targets);

let test_inputs = Matrix::new(2,3, vec![1.5,1.5,1.5,5.1,5.1,5.1]);

model.predict(&test_inputs);

The neural networks are specified via a criterion - similar to Torch. The criterions combine an activation function and a cost function.

You can define your own criterion by implementing the Criterion trait with a concrete ActivationFunc and CostFunc.

Structs

BCECriterion
The binary cross entropy criterion.

MSECriterion
The mean squared error criterion.

NeuralNet
Neural Network struct

Traits

Criterion
Criterion for Neural Networks