Struct rusty_machine::learning::nnet::net_layer::Linear

pub struct Linear { /* private fields */ }

Linear network layer

Represents a fully connected layer with optional bias term

The parameters are a matrix of weights of size I x N where N is the dimensionality of the output and I the dimensionality of the input

Implementations

impl Linear

pub fn new(input_size: usize, output_size: usize) -> Linear

Construct a new Linear layer

pub fn without_bias(input_size: usize, output_size: usize) -> Linear

Construct a Linear layer without a bias term