Module rusty_machine::linalg

The linear algebra module

This module contains reexports of common tools from the rulinalg crate.

Modules

norm

The norm module

Structs

Cholesky

Cholesky decomposition.

FullPivLu

LU decomposition with complete pivoting.

HouseholderComposition

An efficient representation for a composition of Householder transformations.

HouseholderQr

QR decomposition based on Householder reflections.

LUP

Result of unpacking an instance of PartialPivLu.

LUPQ

Result of unpacking an instance of FullPivLu.

Matrix

The Matrix struct.

MatrixSlice

A MatrixSlice

MatrixSliceMut

A mutable MatrixSliceMut

PartialPivLu

LU decomposition with partial pivoting.

QR

The result of unpacking a QR decomposition.

ThinQR

The result of computing a thin (or reduced) QR decomposition.

Vector

The Vector struct.

Enums

Axes

Matrix dimensions

Traits

BaseMatrix

Trait for immutable matrix structs.

BaseMatrixMut

Trait for mutable matrices.

Decomposition

Base trait for decompositions.