Struct rusty_machine::linalg::matrix::MatrixSlice

pub struct MatrixSlice<'a, T: 'a> {
    // some fields omitted
}

A MatrixSlice

This struct provides a slice into a matrix.

The struct contains the upper left point of the slice and the width and height of the slice.

Methods

impl<'a, T> MatrixSlice<'a, T>

fn iter_rows(&self) -> Rows<T>

Iterate over the rows of the matrix slice.

Examples

use rusty_machine::linalg::matrix::{Matrix, MatrixSlice};

let a = Matrix::new(3, 2, (0..6).collect::<Vec<usize>>());
let b = MatrixSlice::from_matrix(&a, [0,0], 2, 2);

// Prints "2" two times.
for row in b.iter_rows() {
    println!("{}", row.len());
}

impl<'a, T> MatrixSlice<'a, T>

fn from_matrix(mat: &'a Matrix<T>, start: [usize; 2], rows: usize, cols: usize) -> MatrixSlice<T>

Produce a matrix slice from a matrix

Examples

use rusty_machine::linalg::matrix::Matrix;
use rusty_machine::linalg::matrix::MatrixSlice;

let a = Matrix::new(3,3, (0..9).collect::<Vec<usize>>());
let slice = MatrixSlice::from_matrix(&a, [1,1], 2, 2);

unsafe fn from_raw_parts(ptr: *const T, rows: usize, cols: usize, row_stride: usize) -> MatrixSlice<'a, T>

Creates a matrix slice from raw parts.

Examples

use rusty_machine::linalg::matrix::MatrixSlice;

let mut a = vec![4.0; 16];

unsafe {
    // Create a matrix slice with 3 rows, and 3 cols
    // The row stride of 4 specifies the distance between the start of each row in the data.
    let b = MatrixSlice::from_raw_parts(a.as_ptr(), 3, 3, 4);
}

Safety

The pointer must be followed by a contiguous slice of data larger than row_stride * rows. If not then other operations will produce undefined behaviour.

Additionally cols should be less than the row_stride. It is possible to use this function safely whilst violating this condition. So long as max(cols, row_stride) * rows is less than the data size.

fn reslice(self, start: [usize; 2], rows: usize, cols: usize) -> MatrixSlice<'a, T>

Produce a matrix slice from an existing matrix slice.

Examples

use rusty_machine::linalg::matrix::Matrix;
use rusty_machine::linalg::matrix::MatrixSlice;

let a = Matrix::new(3,3, (0..9).collect::<Vec<usize>>());
let slice = MatrixSlice::from_matrix(&a, [1,1], 2, 2);
let new_slice = slice.reslice([0,0], 1, 1);

fn iter(&self) -> SliceIter<'a, T>

Returns an iterator over the matrix slice.

Examples

use rusty_machine::linalg::matrix::Matrix;
use rusty_machine::linalg::matrix::MatrixSlice;

let a = Matrix::new(3,3, (0..9).collect::<Vec<usize>>());
let slice = MatrixSlice::from_matrix(&a, [1,1], 2, 2);

let slice_data = slice.iter().map(|v| *v).collect::<Vec<usize>>();
assert_eq!(slice_data, vec![4,5,7,8]);

impl<'a, T: Copy> MatrixSlice<'a, T>

fn into_matrix(self) -> Matrix<T>

Convert the matrix slice into a new Matrix.