Struct rusty_machine::linalg::matrix::MatrixSliceMut

pub struct MatrixSliceMut<T> {
    // some fields omitted
}

A mutable MatrixSlice

This struct provides a mutable slice into a matrix.

The struct contains the upper left point of the slice and the width and height of the slice.

Methods

impl<T> MatrixSliceMut<T>

fn from_matrix(mat: &mut Matrix<T>, start: [usize; 2], rows: usize, cols: usize) -> MatrixSliceMut<T>

Produce a matrix slice from a matrix

Examples

use rusty_machine::linalg::matrix::Matrix;
use rusty_machine::linalg::matrix::MatrixSliceMut;

let mut a = Matrix::new(3,3, (0..9).collect::<Vec<usize>>());
let slice = MatrixSliceMut::from_matrix(&mut a, [1,1], 2, 2);

fn reslice(self, start: [usize; 2], rows: usize, cols: usize) -> MatrixSliceMut<T>

Produce a matrix slice from an existing matrix slice.

Examples

use rusty_machine::linalg::matrix::Matrix;
use rusty_machine::linalg::matrix::MatrixSliceMut;

let mut a = Matrix::new(3,3, (0..9).collect::<Vec<usize>>());
let slice = MatrixSliceMut::from_matrix(&mut a, [1,1], 2, 2);
let new_slice = slice.reslice([0,0], 1, 1);

fn iter(&self) -> SliceIter<T>

Returns an iterator over the matrix slice.

Examples

use rusty_machine::linalg::matrix::Matrix;
use rusty_machine::linalg::matrix::MatrixSliceMut;

let mut a = Matrix::new(3,3, (0..9).collect::<Vec<usize>>());
let slice = MatrixSliceMut::from_matrix(&mut a, [1,1], 2, 2);

let slice_data = slice.iter().map(|v| *v).collect::<Vec<usize>>();
assert_eq!(slice_data, vec![4,5,7,8]);

fn iter_mut(&mut self) -> SliceIterMut<T>

Returns a mutable iterator over the matrix slice.

Examples

use rusty_machine::linalg::matrix::Matrix;
use rusty_machine::linalg::matrix::MatrixSliceMut;

let mut a = Matrix::new(3,3, (0..9).collect::<Vec<usize>>());

{
    let mut slice = MatrixSliceMut::from_matrix(&mut a, [1,1], 2, 2);

    for d in slice.iter_mut() {
        *d = *d + 2;
    }
}

// Only the matrix slice is updated.
assert_eq!(a.into_vec(), vec![0,1,2,3,6,7,6,9,10]);

impl<T: Copy> MatrixSliceMut<T>

fn into_matrix(self) -> Matrix<T>

Convert the matrix slice into a new Matrix.