Struct rusty_machine::linalg::HouseholderComposition

pub struct HouseholderComposition<'a, T>where
    T: 'a,{ /* private fields */ }

An efficient representation for a composition of Householder transformations.

This means that HouseholderComposition represents an operator Q of the form

Q = Q_1 * Q_2 * ... * Q_p

as explained in the documentation for HouseholderQr.

Implementations

impl<'a, T> HouseholderComposition<'a, T>where
T: Float,

pub fn left_multiply_into<X>(&self, matrix: &mut X)where
X: BaseMatrixMut<T>,

Given a matrix A of compatible dimensions, computes the product A <- QA, storing the result in A.

pub fn first_k_columns(&self, k: usize) -> Matrix<T>

Computes the first k columns of the implicitly stored matrix Q.

Panics

k must be less than or equal to m, the number of rows of Q.