Struct rusty_machine::linalg::MatrixSliceMut

pub struct MatrixSliceMut<'a, T> where T: 'a { /* fields omitted */ }

A mutable MatrixSliceMut

This struct provides a mutable slice into a matrix.

The struct contains the upper left point of the slice and the width and height of the slice.

Methods

impl<'a, T> MatrixSliceMut<'a, T>

fn from_matrix(mat: &'a mut Matrix<T>, start: [usize; 2], rows: usize, cols: usize) -> MatrixSliceMut<'a, T>

Produce a MatrixSliceMut from a mutable Matrix

Examples

use rulinalg::matrix::{Matrix, MatrixSliceMut};

let mut a = Matrix::new(3,3, (0..9).collect::<Vec<usize>>());
let slice = MatrixSliceMut::from_matrix(&mut a, [1,1], 2, 2);Run

unsafe fn from_raw_parts(ptr: *mut T, rows: usize, cols: usize, row_stride: usize) -> MatrixSliceMut<'a, T>

Creates a MatrixSliceMut from raw parts.

Examples

use rulinalg::matrix::MatrixSliceMut;

let mut a = vec![4.0; 16];

unsafe {
    // Create a mutable matrix slice with 3 rows, and 3 cols
    // The row stride of 4 specifies the distance between the start of each row in the data.
    let b = MatrixSliceMut::from_raw_parts(a.as_mut_ptr(), 3, 3, 4);
}Run

Safety

The pointer must be followed by a contiguous slice of data larger than row_stride * rows. If not then other operations will produce undefined behaviour.

Additionally cols should be less than the row_stride. It is possible to use this function safely whilst violating this condition. So long as max(cols, row_stride) * rows is less than the data size.

fn reslice(self, start: [usize; 2], rows: usize, cols: usize) -> MatrixSliceMut<'a, T>

Produce a MatrixSliceMut from an existing MatrixSliceMut.

This function will be deprecated. Prefer using BaseMatrixMut::sub_slice_mut instead;

Examples

use rulinalg::matrix::{Matrix, MatrixSliceMut};

let mut a = Matrix::new(3,3, (0..9).collect::<Vec<usize>>());
let slice = MatrixSliceMut::from_matrix(&mut a, [1,1], 2, 2);
let new_slice = slice.reslice([0,0], 1, 1);Run