Struct rusty_machine::stats::dist::gaussian::Gaussian

pub struct Gaussian { /* fields omitted */ }

A Gaussian random variable.

This struct stores both the variance and the standard deviation. This is to minimize the computation required for computing the distribution functions and sampling.

It is most efficient to construct the struct using the from_std_dev constructor.

Methods

impl Gaussian

fn new(mean: f64, variance: f64) -> Gaussian

Creates a new Gaussian random variable from a given mean and variance.

fn from_std_dev(mean: f64, std_dev: f64) -> Gaussian

Creates a new Gaussian random variable from a given mean and standard deviation.