Module rusty_machine::analysis::score

Functions for scoring a set of predictions, i.e. evaluating how close predictions and truth are. All functions in this module obey the convention that higher is better.

Functions

accuracy
Returns the fraction of outputs which match their target.

f1
Returns the f1 score for 2 class classification.

neg_mean_squared_error
Returns the additive inverse of the mean-squared-error of the outputs. So higher is better, and the returned value is always negative.

precision
Returns the precision score for 2 class classification.

recall
Returns the recall score for 2 class classification.

row_accuracy
Returns the fraction of outputs rows which match their target.