Struct rusty_machine::learning::naive_bayes::Gaussian

pub struct Gaussian { /* private fields */ }

The Gaussian Naive Bayes model distribution.

Defines:

p(x|Ck) = ∏i N(xi ; μk, σ2k)

Implementations

impl Gaussian

pub fn theta(&self) -> &Matrix<f64>

Returns the distribution means.

This is a matrix of class by feature means.

pub fn sigma(&self) -> &Matrix<f64>

Returns the distribution variances.

This is a matrix of class by feature variances.