Struct rusty_machine::learning::naive_bayes::NaiveBayes

pub struct NaiveBayes<T: Distribution> { /* fields omitted */ }

The Naive Bayes model.

Implementations

impl<T: Distribution> NaiveBayes<T>

pub fn new() -> NaiveBayes<T>

Create a new NaiveBayes model from a given distribution.

Examples

use rusty_machine::learning::naive_bayes::{NaiveBayes, Gaussian};

// Create a new Gaussian Naive Bayes model.
let _ = NaiveBayes::<Gaussian>::new();

pub fn cluster_count(&self) -> Option<&usize>

Get the cluster count for this model.

Returns an option which is None until the model has been trained.

pub fn class_prior(&self) -> Option<&Vec<f64>>

Get the class prior distribution for this model.

Returns an option which is None until the model has been trained.

pub fn distr(&self) -> Option<&T>

Get the distribution for this model.

Returns an option which is None until the model has been trained.

impl<T: Distribution> NaiveBayes<T>

pub fn get_log_probs(&self, inputs: &Matrix<f64>) -> LearningResult<Matrix<f64>>

Get the log-probabilities per class for each input.