Module rusty_machine::learning::optim::fmincg

Module for the fmincg optimization algorithm.

This algorithm was taken from Andrew Ng’s coursera machine learning course. The function was translated from MATLAB into rust. Original source code can be found here.

The attached license permits use and modification for research and education only.

Copyright (C) 2001 and 2002 by Carl Edward Rasmussen. Date 2002-02-13

(C) Copyright 1999, 2000 & 2001, Carl Edward Rasmussen

Permission is granted for anyone to copy, use, or modify these programs and accompanying documents for purposes of research or education, provided this copyright notice is retained, and note is made of any changes that have been made.

These programs and documents are distributed without any warranty, express or implied. As the programs were written for research purposes only, they have not been tested to the degree that would be advisable in any important application. All use of these programs is entirely at the user’s own risk.

[rusty-machine] Changes made:

Conversion to Rust.

Length hard defaults to the max iterations.

Structs

ConjugateGD

Conjugate Gradient Descent algorithm