Trait rusty_machine::learning::toolkit::kernel::Kernel

pub trait Kernel {
    fn kernel(&self, x1: &[f64], x2: &[f64]) -> f64;
}

The Kernel trait

Requires a function mapping two vectors to a scalar.

Required methods

fn kernel(&self, x1: &[f64], x2: &[f64]) -> f64

The kernel function.

Takes two equal length slices and returns a scalar.

Loading content...