Struct rusty_machine::prelude::Vector

pub struct Vector<T> { /* fields omitted */ }

The Vector struct.

Can be instantiated with any type.

Implementations

impl<T> Vector<T>

pub fn new<U>(data: U) -> Vector<T> where
U: Into<Vec<T>>,

Constructor for Vector struct.

Requires the vector data.

Examples

use rulinalg::vector::Vector;

let vec = Vector::new(vec![1.0,2.0,3.0,4.0]);

pub fn from_fn<F>(size: usize, f: F) -> Vector<T> where
F: FnMut(usize) -> T,

Constructor for Vector struct that takes a function f and constructs a new vector such that V_i = f(i), where i is the index.

Examples

use rulinalg::vector::Vector;

let v = Vector::from_fn(4, |x| x * 3);
assert_eq!(v, vector![0, 3, 6, 9]);

pub fn size(&self) -> usize

Returns the size of the Vector.

pub fn data(&self) -> &Vec<T>

Returns a non-mutable reference to the underlying data.

pub fn mut_data(&mut self) -> &mut [T]

Returns a mutable slice of the underlying data.

pub fn into_vec(self) -> Vec<T>

Consumes the Vector and returns the Vec of data.

pub fn iter(&self) -> Iter<'_, T>

Returns an iterator over the Vector's data.

pub fn iter_mut(&mut self) -> IterMut<'_, T>

Returns an iterator over mutable references to the Vector's data.

pub unsafe fn get_unchecked(&self, index: usize) -> &T

Returns a pointer to the element at the given index, without doing bounds checking.

pub unsafe fn get_unchecked_mut(&mut self, index: usize) -> &mut T

Returns an unsafe mutable pointer to the element at the given index, without doing bounds checking.

impl<T> Vector<T> where
T: Copy,

pub fn apply(self, f: &dyn Fn(T)) -> Vector<T>

Applies a function to each element in the vector.

Examples

use rulinalg::vector::Vector;
fn add_two(a: f64) -> f64 {
    a + 2f64
}

let a = vector![0.; 4];

let b = a.apply(&add_two);

assert_eq!(b, vector![2.0; 4]);

impl<T> Vector<T> where
T: PartialOrd<T> + Copy,

pub fn argmax(&self) -> (usize, T)

Find the argmax of the Vector.

Returns the index of the largest value in the vector.

Examples

use rulinalg::vector::Vector;

let a = vector![1.0,2.0,0.0,5.0];
let b = a.argmax();
assert_eq!(b.0, 3);
assert_eq!(b.1, 5.0);

pub fn argmin(&self) -> (usize, T)

Find the argmin of the Vector.

Returns the index of the smallest value in the vector.

Examples

use rulinalg::vector::Vector;

let a = vector![1.0, 2.0, 0.0, 5.0];
let b = a.argmin();
assert_eq!(b.0, 2);
assert_eq!(b.1, 0.0);

pub fn select(&self, idxs: &[usize]) -> Vector<T>

Select elements from the Vector and form a new Vector from them.

Examples

use rulinalg::vector::Vector;

let a = vector![1.0, 2.0, 3.0, 4.0, 5.0];

let a_lower = a.select(&[2, 3, 4]);

// Prints [3,4,5]
assert_eq!(a_lower, vector![3.0, 4.0, 5.0]);

impl<T> Vector<T> where
T: Clone + Zero,

pub fn zeros(size: usize) -> Vector<T>

Constructs Vector of all zeros.

Requires the size of the vector.

Examples

use rulinalg::vector::Vector;

let vec = Vector::<f64>::zeros(10);

impl<T> Vector<T> where
T: Clone + One,

pub fn ones(size: usize) -> Vector<T>

Constructs Vector of all ones.

Requires the size of the vector.

Examples

use rulinalg::vector::Vector;

let vec = Vector::<f64>::ones(10);

impl<T> Vector<T> where
T: Add<T, Output = T> + Mul<T, Output = T> + Zero + Copy,

pub fn dot(&self, v: &Vector<T>) -> T

Compute dot product with specified Vector.

Examples

use rulinalg::vector::Vector;

let a = vector![1.0, 2.0, 3.0, 4.0];
let b = vector![2.0; 4];

let c = a.dot(&b);
assert_eq!(c, 20.0);

impl<T> Vector<T> where
T: Copy + Add<T, Output = T> + Zero,

pub fn sum(&self) -> T

The sum of the vector.

Returns the sum of all elements in the vector.

Examples

use rulinalg::vector::Vector;

let a = vector![1.0, 2.0, 3.0, 4.0];

let c = a.sum();
assert_eq!(c, 10.0);

impl<T> Vector<T> where
T: Copy + Mul<T, Output = T>,

pub fn elemul(&self, v: &Vector<T>) -> Vector<T>

The elementwise product of two vectors.

Examples

use rulinalg::vector::Vector;

let a = vector![1.0, 2.0, 3.0, 4.0];
let b = vector![1.0, 2.0, 3.0, 4.0];

let c = &a.elemul(&b);
assert_eq!(c, &vector![1.0, 4.0, 9.0, 16.0]);

impl<T> Vector<T> where
T: Copy + Div<T, Output = T>,

pub fn elediv(&self, v: &Vector<T>) -> Vector<T>

The elementwise division of two vectors.

Examples

use rulinalg::vector::Vector;

let a = vector![1.0, 2.0, 3.0, 4.0];
let b = vector![1.0, 2.0, 3.0, 4.0];

let c = &a.elediv(&b);
assert_eq!(c, &vector![1.0; 4]);

impl<T> Vector<T> where
T: Float,

pub fn norm<N>(&self, norm: N) -> T where
N: VectorNorm<T>,

Compute vector norm for vector.

Examples

use rulinalg::vector::Vector;
use rulinalg::norm::Euclidean;

let a = vector![3.0, 4.0];
let c = a.norm(Euclidean);

assert_eq!(c, 5.0);

pub fn metric<M>(&self, v: &Vector<T>, m: M) -> T where
M: VectorMetric<T>,

Compute metric distance between two vectors.

Examples

use rulinalg::vector::Vector;
use rulinalg::norm::Euclidean;

let a = vector![3.0, 4.0];
let b = vector![0.0, 8.0];

// Compute the square root of the sum of
// elementwise squared-differences
let c = a.metric(&b, Euclidean);
assert_eq!(c, 5.0);

impl<T> Vector<T> where
T: FromPrimitive + Float,

pub fn mean(&self) -> T

The mean of the vector.

Returns the arithmetic mean of the vector.

Examples

use rulinalg::vector::Vector;

let a = vector![1.0, 2.0, 3.0, 4.0];

let c = a.mean();
assert_eq!(c, 2.5);

pub fn variance(&self) -> T

The variance of the vector.

Returns the unbiased sample variance of the vector.

Examples

use rulinalg::vector::Vector;

let a = vector![1.0, 2.0, 3.0, 4.0];

let c = a.variance();
assert_eq!(c, 5.0 / 3.0);